requirements.txt
asgiref==3.2.9
dj-database-url==0.5.0
Django==3.0.7
django-guardian==2.3.0
django-multi-email-field==0.6.1
django-widget-tweaks==1.4.8
gunicorn==20.0.4
Pillow==7.1.2
psycopg2-binary==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
